fix code errors

+++ Assets/Scripts/Trail Making/TrailLevel1.cs
@@ -295,7 +295,7 @@ public class TrailLevel1 : MonoBehaviour
             mistakesIndicator[i].GetComponent<Image>().color = Color.green;
         }
     }
-    float CalculateAccuracy(int correct, int total)
+    float CalculateAccuracy(float correct, float total)
     {
         float acc = correct / total;
         return acc;
