+++ Assets/Scripts/Trail Making/TrailLevel3.cs
@@ -279,7 +279,7 @@ public class TrailLevel3 : MonoBehaviour
             mistakesIndicator_3[i].GetComponent<Image>().color = Color.green;
         }
     }
-    float CalculateAccuracy(int correct, int total)
+    float CalculateAccuracy(float correct, float total)
     {
         float acc = correct / total;
         return acc;
